pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mcp-robot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "paramiko>=3.4",
    "google-genai>=1.0",
    "Pillow>=10.4",
]

[project.scripts]
mcp-robot = "mcp_robot.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_robot*"]
